gnome/generate_gir: place all --extra-library and -L arguments after the --library
This fixes the ordering of the resulting compile/link command used by
g-ir-scanner in the following scenario:

1. There exists an already built copy of the shared library that is
   being scanned inside the install prefix.
2. There may be an internal dependency that has been built by this
   project which has a dependency on another library inside the install
   prefix.
3. The result is that there will be -L linker arguments added for the
   install prefix before the -L arguments for the library that is being
   scanned meaning that the old library from the install would be linked
   instead of the required newly built library.

This is only a partial fix as there may be a scenario where there may be
multiple built dependencies in this subproject that may depend on a
library inside the install prefix.  If any of said built libraries were
also previously installed in the install dir, those old libraries may be
used instead in the same manner.